feat: fetch matrix pill info
