+++ src/models/mod.rs
@@ -3,6 +3,7 @@
 pub(crate) mod async_requests;
 pub mod event_bridge;
 pub mod events;
+pub mod matrix_uri;
 pub mod misc;
 pub mod profile;
 pub(crate) mod room_display_name;
