+++ src/utils.rs
@@ -1,12 +1,9 @@
-use matrix_sdk::ruma::matrix_uri::MatrixId;
-use matrix_sdk::{IdParseError, OwnedServerName};
-use serde::Serialize;
 use std::borrow::Cow;
 use tokio::sync::{broadcast, mpsc};
 use tokio::time::{Duration, sleep};
 use tracing::warn;
 
-use matrix_sdk::ruma::{MatrixToUri, MatrixUri, OwnedRoomOrAliasId, OwnedUserId, RoomId};
+use matrix_sdk::ruma::RoomId;
 use matrix_sdk_ui::timeline::{EventTimelineItem, TimelineDetails};
 
 use crate::events::timeline::TimelineKind;
@@ -232,66 +229,4 @@ pub(crate) enum VecDiff<T> {
     PopBack,
     /// Truncate the list to the given length.
     Truncate { length: usize },
-}
-
-/// Tries to extract a room address (Alias or ID) from the given text.
-///
-/// This function is quite flexible and will attempt to parse `text` as:
-/// * A Room ID (with a leading `!`).
-/// * A Room Alias (with a leading `#`).
-/// * A `https://matrix.to` URI, which includes either a room alias, or a room ID plus `via` servers.
-/// * A `matrix:` scheme URI, which is similar to above.
-pub(crate) fn parse_address(
-    text: &str,
-) -> Result<(OwnedRoomOrAliasId, Vec<OwnedServerName>), IdParseError> {
-    match OwnedRoomOrAliasId::try_from(text) {
-        Ok(room_or_alias_id) => Ok((room_or_alias_id, Vec::new())),
-        Err(e) => {
-            let uri_result = MatrixToUri::parse(text)
-                .map(|uri| (uri.id().clone(), uri.via().to_owned()))
-                .or_else(|_| {
-                    MatrixUri::parse(text).map(|uri| (uri.id().clone(), uri.via().to_owned()))
-                });
-
-            if let Ok((matrix_id, via)) = uri_result
-                && let Some(room_or_alias_id) = match matrix_id {
-                    MatrixId::Room(room_id) => Some(room_id.into()),
-                    MatrixId::RoomAlias(alias) => Some(alias.into()),
-                    MatrixId::Event(room_or_alias_id, _) => Some(room_or_alias_id),
-                    _ => None,
-                }
-            {
-                return Ok((room_or_alias_id, via));
-            }
-            Err(e)
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(
-    rename_all = "camelCase",
-    rename_all_fields = "camelCase",
-    tag = "kind",
-    content = "payload"
-)]
-pub enum MatrixUriIntent {
-    Room((OwnedRoomOrAliasId, Vec<OwnedServerName>)),
-    User(OwnedUserId),
-}
-
-pub fn get_matrix_uri_intent(text: &str) -> Result<MatrixUriIntent, IdParseError> {
-    let uri = MatrixUri::parse(text)?;
-    match uri.id().clone() {
-        MatrixId::Room(room_id) => Ok(MatrixUriIntent::Room((room_id.into(), uri.via().to_vec()))),
-        MatrixId::RoomAlias(alias) => Ok(MatrixUriIntent::Room((alias.into(), uri.via().to_vec()))),
-        MatrixId::Event(room_or_alias_id, _) => Ok(MatrixUriIntent::Room((
-            room_or_alias_id,
-            uri.via().to_vec(),
-        ))),
-        MatrixId::User(user_id) => Ok(MatrixUriIntent::User(user_id)),
-        _ => Err(IdParseError::InvalidMatrixUri(
-            matrix_sdk::ruma::MatrixUriError::UnknownQueryItem,
-        )),
-    }
 }
